// frontend/src/client/types.gen.ts
export type InvoicePublic = {
    invoice_number: string;
    invoice_code: string;
    amount: number;
    invoice_date?: (string | null);
    seller?: (string | null);
    buyer?: (string | null);
    tax_amount?: (number | null);
    remark?: (string | null);
    id: string;
    contract_id: string;
    file_path?: (string | null);
    parsed_data?: ({
    [key: string]: unknown;
} | null);
    created_at?: (string | null);
};